fix(a2-audit): Paso 1 - migrar a integer 13 casos + actualizar tests
FASE A2 - PASO 1: Migración de 143 casts (float) residuales.

CAMBIOS EN CÓDIGO (13 cambios en 6 archivos):

📦 ACCOUNTING (4):
   - JournalEntry.totalDebits(): float → int
   - JournalEntry.totalCredits(): float → int
   - Cast sum('debit_amount') y sum('credit_amount') a int

📦 SHARED (1):
   - MoneyFormatter.parse(): tipo de retorno int|float (union type)
   - Lógica condicional: retorna int si decimals=0, float si decimals>0
   - Preserva semántica correcta según ADR-018

📦 KITCHEN (1):
   - BroadcastOrderPaid: 'total' enviado como int

📦 PAYMENTS (7):
   - CashSessionResource: difference como int
   - CashSessionController: openingAmount y closingAmount como int
   - CashSession entity: cashSales, cashTips, brutExpected, paidOut como int

NO TOCADOS (float legítimo, 4 casos):
   - Tax.rate (3): tasas de impuesto (ej: 19.0)
   - MoneyFormatter.format(): number_format requiere float para display

ACTUALIZACIONES DE TESTS (3 archivos):

📝 MoneyFormatterTest.php:
   - Tests de CLP ahora tienen configuración explícita decimals=0
   - parse('12.990') → retorna 12990 (int) para CLP
   - parse('$ 12.990') → retorna 12990 (int) para CLP
   - parse('1,234.56') → retorna 1234.56 (float) para USD (sin cambio)

📝 LedgerTest.php:
   - totalDebits()/totalCredits() → retornan 11900 (int)

📝 BroadcastEventsTest.php:
   - BroadcastOrderPaid 'total' → 11900 (int)

RESULTADO:
✅ 1021 tests passing, 0 failed
✅ 3050 assertions
✅ Consistencia con ADR-018 (CLP integer end-to-end)

PRÓXIMO PASO: Paso 2 (Catalog + Orders, ~20 casos)

+++ app/Modules/Accounting/Domain/Entities/JournalEntry.php
@@ -73,17 +73,17 @@ class JournalEntry extends Model
     /**
      * Calcula el total de débitos del asiento.
      */
-    public function totalDebits(): float
+    public function totalDebits(): int  // ADR-018: CLP entero
     {
-        return (float) $this->ledgerEntries()->sum('debit_amount');
+        return (int) $this->ledgerEntries()->sum('debit_amount');  // ADR-018
     }
 
     /**
      * Calcula el total de créditos del asiento.
      */
-    public function totalCredits(): float
+    public function totalCredits(): int  // ADR-018: CLP entero
     {
-        return (float) $this->ledgerEntries()->sum('credit_amount');
+        return (int) $this->ledgerEntries()->sum('credit_amount');  // ADR-018
     }
 
     public function scopeByReference($query, ReferenceType $type, int $referenceId)
